+++ app/src/main/java/com/humberto/tasky/agenda/presentation/agenda_details/AgendaDetailsScreen.kt
@@ -3,7 +3,6 @@ package com.humberto.tasky.agenda.presentation.agenda_details
 import android.widget.Toast
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -35,7 +34,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
@@ -46,7 +44,7 @@ import com.humberto.tasky.R
 import com.humberto.tasky.agenda.presentation.agenda_details.components.AddEventAttendeeDialog
 import com.humberto.tasky.agenda.presentation.agenda_details.components.AgendaItemIndicator
 import com.humberto.tasky.agenda.presentation.agenda_details.components.AttendanceFilter
-import com.humberto.tasky.agenda.presentation.agenda_details.components.AttendeeListItem
+import com.humberto.tasky.agenda.presentation.agenda_details.components.AttendeeList
 import com.humberto.tasky.agenda.presentation.agenda_details.components.DateTimeSection
 import com.humberto.tasky.agenda.presentation.agenda_details.components.ManageAgendaItemStateButton
 import com.humberto.tasky.agenda.presentation.agenda_details.components.PhotosSection
@@ -343,11 +341,13 @@ private fun AgendaDetailsScreen(
                 color = MaterialTheme.colorScheme.tertiary
             )
             if(agendaItem is AgendaItemDetails.Event) {
-                AddEventAttendeeDialog(
-                    onDismiss = { onAction(AgendaDetailsAction.OnDismissAttendeeDialog) },
-                    onAdd = { onAction(AgendaDetailsAction.OnAddAttendeeClick) },
-                    agendaItem = agendaItem
-                )
+                if(agendaItem.isAddingAttendee) {
+                    AddEventAttendeeDialog(
+                        onDismiss = { onAction(AgendaDetailsAction.OnDismissAttendeeDialog) },
+                        onAdd = { onAction(AgendaDetailsAction.OnAddAttendeeClick) },
+                        agendaItem = agendaItem
+                    )
+                }
                 Column(
                     modifier = Modifier
                         .padding(bottom = 32.dp)
@@ -401,55 +401,6 @@ private fun AgendaDetailsScreen(
                 modifier = Modifier,
                 agendaItem = agendaItem
             )
-        }
-    }
-}
-
-@Composable
-fun AttendeeList(
-    modifier: Modifier = Modifier,
-    attendeeList: List<AttendeeUi>,
-    selectedAttendanceFilter: FilterType
-) {
-    val goingList = attendeeList.filter { it.isGoing }
-    val notGoingList = attendeeList.filter { !it.isGoing }
-
-    Column(
-        modifier = modifier.fillMaxWidth(),
-        verticalArrangement = Arrangement.spacedBy(4.dp)
-    ) {
-        if (selectedAttendanceFilter == FilterType.ALL || selectedAttendanceFilter == FilterType.GOING) {
-            if (goingList.isNotEmpty()) {
-                Text(
-                    text = stringResource(id = R.string.filter_going),
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    fontWeight = FontWeight.Medium,
-                    modifier = Modifier.padding(vertical = 8.dp)
-                )
-                goingList.forEach { attendee ->
-                    AttendeeListItem(
-                        onDeleteClick = { },
-                        attendee = attendee
-                    )
-                }
-            }
-        }
-
-        if (selectedAttendanceFilter == FilterType.ALL || selectedAttendanceFilter == FilterType.NOT_GOING) {
-            if (notGoingList.isNotEmpty()) {
-                Text(
-                    text = stringResource(id = R.string.filter_not_going),
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    fontWeight = FontWeight.Medium,
-                    modifier = Modifier.padding(vertical = 8.dp)
-                )
-                notGoingList.forEach { attendee ->
-                    AttendeeListItem(
-                        onDeleteClick = { },
-                        attendee = attendee
-                    )
-                }
-            }
         }
     }
 }
